Required changes code refactoring minor improvements
