+++ ToyBox/Classes/MonkeyPatchin/BagOfPatches/MiscRT.cs
@@ -281,16 +281,16 @@ namespace ToyBox.BagOfPatches {
         [HarmonyPatch(typeof(Polymorph), nameof(Polymorph.TryReplaceView))]
         private static class Polymorph_TryReplaceView_Patch {
             private static void Postfix(Polymorph __instance) {
-                Vector3 scale = PartyEditor.lastScaleSize.GetValueOrDefault(__instance.Owner.HashKey(), new Vector3(1, 1, 1));
-                __instance.Owner.View.transform.localScale = scale;
+                float scale = PartyEditor.lastScaleSize.GetValueOrDefault(__instance.Owner.HashKey(), 1);
+                __instance.Owner.View.transform.localScale = new Vector3(scale, scale, scale);
             }
         }
 
         [HarmonyPatch(typeof(Polymorph), nameof(Polymorph.RestoreView))]
         private static class Polymorph_RestoreView_Patch {
             private static void Postfix(Polymorph __instance) {
-                Vector3 scale = PartyEditor.lastScaleSize.GetValueOrDefault(__instance.Owner.HashKey(), new Vector3(1, 1, 1));
-                __instance.Owner.View.transform.localScale = scale;
+                float scale = PartyEditor.lastScaleSize.GetValueOrDefault(__instance.Owner.HashKey(), 1);
+                __instance.Owner.View.transform.localScale = new Vector3(scale, scale, scale);
             }
         }
 
@@ -308,10 +308,10 @@ namespace ToyBox.BagOfPatches {
 
                     foreach (BaseUnitEntity cha in Game.Instance.State.AllUnits.Where((u) => u.HashKey().Equals(ID))) {
 
-                        var scale = Main.Settings.perSave.characterModelSizeMultiplier.GetValueOrDefault(ID, new(1, 1, 1));
+                        float scale = Main.Settings.perSave.characterModelSizeMultiplier.GetValueOrDefault(ID, 1);
 
-                        cha.View.gameObject.transform.localScale = new Vector3(scale.Item1, scale.Item2, scale.Item3);
-                        PartyEditor.lastScaleSize[cha.HashKey()] = new Vector3(scale.Item1, scale.Item2, scale.Item3);
+                        cha.View.gameObject.transform.localScale = new Vector3(scale, scale, scale);
+                        PartyEditor.lastScaleSize[cha.HashKey()] = scale;
                     }
                 }
 
@@ -319,15 +319,19 @@ namespace ToyBox.BagOfPatches {
 
                     foreach (BaseUnitEntity cha in Game.Instance.State.AllUnits.Where((u) => u.HashKey().Equals(ID))) {
 
-                        PartyEditor.skeletonReplacers[cha.HashKey()] = new PartyEditor.CharacterSkeletonReplacer(cha);
+                        PartyEditor.skeletonReplacers[cha.HashKey()] = new SkeletonReplacer(cha);
                         PartyEditor.skeletonReplacers[cha.HashKey()].ApplyBonesModification(cha, true);
                     }
                 }
 
                 foreach (var ID in Main.Settings.perSave.characterSizeModifier.Keys) {
+
                     foreach (BaseUnitEntity cha in Game.Instance.State.AllUnits.Where((u) => u.HashKey().Equals(ID))) {
+
                         Kingmaker.Enums.Size size;
+
                         if (Main.Settings.perSave.characterSizeModifier.TryGetValue(ID, out size)) {
+
                             cha.Descriptor().State.Size = size;
                         }
                     }
