+++ ToyBox/Localization/en.json
@@ -625,18 +625,16 @@
         "Wrist": "Wrist",
         "You": "You",
         "You can enable hot keys to teleport members of your party to your mouse cursor on Area or the Global Map": "You can enable hot keys to teleport members of your party to your mouse cursor on Area or the Global Map",
-        "Visual character size multiplier X": "Visual character size multiplier X",
-        "Visual character size multiplier Y": "Visual character size multiplier Y",
-        "Visual character size multiplier Z": "Visual character size multiplier Z",
         "Body parts offsets and position": "Offsets and positions of body parts",
         "Body parts global scales": "Global scales of body parts",
         "Body parts local sizes": "Local sizes of body parts",
         "Equipment elements sizes": "Sizes of equipment elements",
         "OF_position": "Character Z-axis position",
+        "OF_backpack_and_cloakX": "Backpack and cloak X-axis offset",
+        "OF_backpack_and_cloakZ": "Backpack and cloak Z-axis offset",
         "OF_shouldersX": "Shoulder X-axis offset",
         "OF_shouldersZ": "Shoulder Z-axis offset",
         "OF_upper_arms": "Upper arm offset",
-        "OF_fore_arms": "Forearm offset",
         "OF_upper_legs": "Upper leg offset",
         "SC_pelvisX": "Pelvis bone scale X-axis",
         "SC_pelvisY": "Pelvis bone scale Y-axis",
