+++ ToyBox/Localization/ru-RU.json
@@ -625,18 +625,16 @@
         "Wrist": "Запястье",
         "You": "Вы",
         "You can enable hot keys to teleport members of your party to your mouse cursor on Area or the Global Map": "Вы можете включить горячие клавиши для телепортации членов вашей Партии к указателю мыши на Локации или Глобальной карте",
-        "Visual character size multiplier X": "Множитель визуального размера X",
-        "Visual character size multiplier Y": "Множитель визуального размера Y",
-        "Visual character size multiplier Z": "Множитель визуального размера Z",
         "Body parts offsets and position": "Смещения и положение частей тела",
         "Body parts global scales": "Глобальные масштабы частей тела",
         "Body parts local sizes": "Локальные размеры частей тела",
         "Equipment elements sizes": "Размеры элементов экипировки",
         "OF_position": "Положение персонажа по оси Z",
+        "OF_backpack_and_cloakX": "Смещение рюкзака и плаща оси X",
+        "OF_backpack_and_cloakZ": "Смещение рюкзака и плаща оси Z",
         "OF_shouldersX": "Смещение плеч по оси X",
         "OF_shouldersZ": "Смещение плеч по оси Z",
         "OF_upper_arms": "Смещение верхних рук",
-        "OF_fore_arms": "Смещение предплечий",
         "OF_upper_legs": "Смещение верхних ног",
         "SC_pelvisX": "Масштаб первой кости тела по X",
         "SC_pelvisY": "Масштаб первой кости тела по Y",
